refactor: move web search routes around
