+++ backend/api/src/http/endpoints/image.rs
@@ -142,41 +142,6 @@ pub mod user {
     }
 }
 
-mod web_library {
-    use core::settings::RuntimeSettings;
-
-    use paperclip::actix::{
-        api_v2_operation,
-        web::{Data, Json, Query},
-    };
-
-    use shared::{
-        api::{endpoints, ApiEndpoint},
-        domain::image::web::WebImageSearchResponse,
-    };
-
-    use crate::{error, extractor::WrapAuthClaimsNoDb};
-
-    /// Search for images in the web image library.
-    #[api_v2_operation]
-    pub async fn search(
-        runtime_settings: Data<RuntimeSettings>,
-        _claims: WrapAuthClaimsNoDb,
-        query: Query<<endpoints::image::web::Search as ApiEndpoint>::Req>,
-    ) -> Result<Json<<endpoints::image::web::Search as ApiEndpoint>::Res>, error::Server> {
-        let query = query.into_inner();
-
-        // todo: handle empty queries (they're invalid in bing)
-
-        let res = match &runtime_settings.bing_search_key {
-            Some(key) => crate::image_search::get_images(&query.q, key).await?,
-            None => WebImageSearchResponse { images: Vec::new() },
-        };
-
-        Ok(Json(res))
-    }
-}
-
 // attempts to grab a uuid out of a string in the shape:
 // Key (<key>)=(<uuid>)<postfix>
 fn extract_uuid(s: &str) -> Option<Uuid> {
@@ -455,11 +420,5 @@ pub fn configure(cfg: &mut ServiceConfig<'_>) {
     .route(
         image::user::List::PATH,
         image::user::List::METHOD.route().to(self::user::list),
-    )
-    .route(
-        image::web::Search::PATH,
-        image::web::Search::METHOD
-            .route()
-            .to(self::web_library::search),
     );
 }
