+++ shared/rust/src/domain/image.rs
@@ -47,40 +47,6 @@ pub mod user {
     }
 }
 
-/// Types for the web image library.
-pub mod web {
-    #[cfg(feature = "backend")]
-    use paperclip::actix::Apiv2Schema;
-    use serde::{Deserialize, Serialize};
-
-    /// Search for images via the given query string.
-    #[derive(Serialize, Deserialize, Clone, Debug)]
-    #[cfg_attr(feature = "backend", derive(Apiv2Schema))]
-    pub struct WebImageSearchQuery {
-        /// The query string.
-        pub q: String,
-    }
-
-    /// A single image as returned from the web
-    #[derive(Serialize, Deserialize, Debug)]
-    #[cfg_attr(feature = "backend", derive(Apiv2Schema))]
-    pub struct WebImageSearchItem {
-        /// A URL to the thumbnail of the image.
-        pub thumbnail_url: String,
-        /// A URL to the original image.
-        pub url: String,
-    }
-
-    /// Response for successful search.
-    /// TODO: support pagation
-    #[derive(Serialize, Deserialize, Debug)]
-    #[cfg_attr(feature = "backend", derive(Apiv2Schema))]
-    pub struct WebImageSearchResponse {
-        /// the images returned.
-        pub images: Vec<WebImageSearchItem>,
-    }
-}
-
 /// Represents different kinds of images (which affects how the size is stored in the db)
 #[derive(Serialize, Deserialize, Copy, Clone, Debug)]
 #[cfg_attr(feature = "backend", derive(sqlx::Type))]
